feat: hieuhq - implement cart

+++ src/app/client/main/cart/cart.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 import { CartService } from 'src/app/shared/services/cart.service';
 import { environment } from 'src/environments/environment';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-cart',
@@ -39,12 +40,40 @@ export class CartComponent implements OnInit {
   }
 
   removeItemFromCart(id: string) {
-    this.cartService
-      .deleteItemFromCart(id)
-      .subscribe((data: any) => {
-        console.log("Remove ok");
-        this.getCarts();
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton:
+          'btn btn-success',
+        cancelButton:
+          'btn btn-default',
+      },
+      buttonsStyling: false,
+    });
+
+    swalWithBootstrapButtons
+      .fire({
+        text: 'Xác nhận xóa khoá học này?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Xác nhận',
+        cancelButtonText: 'Hủy',
+        reverseButtons: true,
+      })
+      .then((result) => {
+        if (result.isConfirmed) {
+          this.cartService
+            .deleteItemFromCart(id)
+            .subscribe((data: any) => {
+              this.getCarts();
+            });
+
+          swalWithBootstrapButtons.fire('Xóa thành công!');
+        } else if (result.dismiss === Swal.DismissReason.cancel) {
+          swalWithBootstrapButtons.fire('Hủy thành công!');
+        }
       });
+
+
   }
 
 }
